Introduce settable default step for TF 2.0 summary API (experimental)
This adds tf.summary.experimental.set_step() and tf.summary.experimental.get_step() functions to the TF 2.0 summary API. Their purpose is to allow setting a default value for the `step` parameter that had up until now been required on the new summary API writing functions like scalar(), image(), etc. Those parameters are now optional, though if there is no default step configured an exception will still be raised.

The step can be manually set to a constant value in each iteration of the train loop, or you can set the step to a tf.Variable and then update the variable independently. Only the latter works well across a tf.function boundary because constant values will be captured when the function is traced, so updates outside the function won't be reflected inside it.

PiperOrigin-RevId: 235675144