Use IS_MOBILE_PLATFORM instead of __ANDROID__ for C API
This changes the C API to test for the presence of the IS_MOBILE_PLATFORM or
IS_SLIM_BUILD macro symbol. This allows certain features that are not supported
on mobile (e.g. distributed runtime) to be excluded on platforms other than
Android.

This change is a prerequisite for including the TensorFlow C API on mobile
platforms.

PiperOrigin-RevId: 236174151